Activity score 0-10 scale + Room version bump to 12
Changed activity score from 0-100 integer to 0.0-10.0 float.
Weights: votes 4.0, questions 2.0, speeches 2.0, committees 2.0.
Updated MpStatsEntity, MpStats, ActivityScoreCalculator, UI components.
Bumped Room version 11→12, CURRENT_SEED_VERSION 2→3.

+++ core/data/src/main/java/com/goveye/app/data/local/entity/MpStatsEntity.kt
@@ -13,7 +13,7 @@ import androidx.room.PrimaryKey
  * - questionCount, speechCount, committeeCount: COUNT queries
  * - voteParticipationRate: voted divisions / total divisions in house
  * - rebellionRate, rebellionCount, totalDivisionsVoted: party-majority method
- * - activityScore: weighted score (votes 40%, questions 20%, speeches 20%, committees 20%)
+ * - activityScore: weighted score 0.0-10.0 (votes 4.0, questions 2.0, speeches 2.0, committees 2.0)
  * - percentiles: rank-based percentile across same-house peers
  */
 @Entity(tableName = "mp_stats")
@@ -27,7 +27,7 @@ data class MpStatsEntity(
     val rebellionRate: Float,
     val rebellionCount: Int,
     val totalDivisionsVoted: Int,
-    val activityScore: Int,
+    val activityScore: Float,
     val rebellionPercentile: Int,
     val participationPercentile: Int,
     val questionsPercentile: Int,
